Move Jira functionality into jira package
