+++ jira/comment/comment.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package comments
+package comment
 
 import (
 	"fmt"
